Update to upstream version '4.0.12'
with Debian dir 7faca2f34003d4a6eb54630255ad192b185a93b3

+++ src/mongo/util/procparser.cpp
@@ -1,29 +1,31 @@
+
 /**
- * Copyright (C) 2016 MongoDB Inc.
+ *    Copyright (C) 2018-present MongoDB, Inc.
  *
- * This program is free software: you can redistribute it and/or  modify
- * it under the terms of the GNU Affero General Public License, version 3,
- * as published by the Free Software Foundation.
+ *    This program is free software: you can redistribute it and/or modify
+ *    it under the terms of the Server Side Public License, version 1,
+ *    as published by MongoDB, Inc.
  *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
+ *    This program is distributed in the hope that it will be useful,
+ *    but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *    Server Side Public License for more details.
  *
- * You should have received a copy of the GNU Affero General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ *    You should have received a copy of the Server Side Public License
+ *    along with this program. If not, see
+ *    <http://www.mongodb.com/licensing/server-side-public-license>.
  *
- * As a special exception, the copyright holders give permission to link the
- * code of portions of this program with the OpenSSL library under certain
- * conditions as described in each individual source file and distribute
- * linked combinations including the program with the OpenSSL library. You
- * must comply with the GNU Affero General Public License in all respects
- * for all of the code used other than as permitted herein. If you modify
- * file(s) with this exception, you may extend this exception to your
- * version of the file(s), but you are not obligated to do so. If you do not
- * wish to do so, delete this exception statement from your version. If you
- * delete this exception statement from all source files in the program,
- * then also delete it in the license file.
+ *    As a special exception, the copyright holders give permission to link the
+ *    code of portions of this program with the OpenSSL library under certain
+ *    conditions as described in each individual source file and distribute
+ *    linked combinations including the program with the OpenSSL library. You
+ *    must comply with the Server Side Public License in all respects for
+ *    all of the code used other than as permitted herein. If you modify file(s)
+ *    with this exception, you may extend this exception to your version of the
+ *    file(s), but you are not obligated to do so. If you do not wish to do so,
+ *    delete this exception statement from your version. If you delete this
+ *    exception statement from all source files in the program, then also delete
+ *    it in the license file.
  */
 
 #define MONGO_LOG_DEFAULT_COMPONENT ::mongo::logger::LogComponent::kFTDC
@@ -278,7 +280,7 @@ Status parseProcStat(const std::vector<StringData>& keys,
                     value = 0;
                 }
 
-                builder->appendNumber(key, value);
+                builder->appendNumber(key, static_cast<long long>(value));
             }
         }
     }
@@ -383,10 +385,10 @@ Status parseProcMemInfo(const std::vector<StringData>& keys,
                     keyWithSuffix.append("_kb");
                 }
 
-                builder->appendNumber(keyWithSuffix, value);
+                builder->appendNumber(keyWithSuffix, static_cast<long long>(value));
             } else {
 
-                builder->appendNumber(key, value);
+                builder->appendNumber(key, static_cast<long long>(value));
             }
         }
     }
@@ -478,7 +480,8 @@ Status parseProcNetstat(const std::vector<StringData>& keys,
                     StringData stringValue((*valuesIt).begin(), (*valuesIt).end());
                     uint64_t value;
                     if (parseNumberFromString(stringValue, &value).isOK()) {
-                        builder->appendNumber(prefix.toString() + key.toString(), value);
+                        builder->appendNumber(prefix.toString() + key.toString(),
+                                              static_cast<long long>(value));
                         foundKeys = true;
                     }
                 }
@@ -614,7 +617,7 @@ Status parseProcDiskStats(const std::vector<StringData>& disks,
                 BSONObjBuilder sub(builder->subobjStart(disk));
 
                 for (size_t index = 0; index < stats.size() && index < kDiskFieldCount; ++index) {
-                    sub.appendNumber(kDiskFields[index], stats[index]);
+                    sub.appendNumber(kDiskFields[index], static_cast<long long>(stats[index]));
                 }
 
                 sub.doneFast();
